// src/app/cyprus/types.ts
export interface HotelInfo {
  name?: string
  address?: string
  images?: Array<{ original: string }>
  meal_type?: string
}

export interface FlightInfo {
  airline?: string
  flightNumber?: string
  flightNo?: string
  classType?: string
  cabinClass?: string
  departureTime: string
  arrivalTime: string
  departureDate?: string
  arrivalDate?: string
  origin: string
  destination: string
  connectionCount?: number
  connections?: string[]
}
export interface TransferInfo {
  vehicleName?: string
  transferTitle?: string
  pickupLocationName?: string
  dropLocationName?: string
}
export interface CyprusSummaryProps {
  hotel?: HotelInfo
  checkInDate: Date | string
  checkOutDate: Date | string
  adults?: number
  children?: number
  departureFlight?: FlightInfo
  returnFlight?: FlightInfo
  departureTransfer?: TransferInfo
  returnTransfer?: TransferInfo
  showCancelWarranty?: boolean
  useCancelWarranty?: boolean
  title?: string
  isFlight?: boolean
  isTransfer?: boolean
}